ELEMENT_INFO: Implement contexts
The same element in the parsed XML can appear in multiple contexts,
requiring multiple descriptions how to handle it by the parser.

This commit introduces the ability to define a context (identified by a
symbol) for an element and ability to describe element parsing in
different contexts.

This commit also makes <list> define the :list context and specifies how
<element> behaves in it. This is needed because <map> (parsing of which
will be implemented in the next commit) can also contain <element>, but
with a different structure & semantics.