Save the filename in AST
This commit makes it so that the AST root node will always contain a
|filename| property with the translated .ycp file name. If the file name
is unknown (e.g. because bin/y2r translates a standard input),
"default.yml" is used.

The file name is needed because when we compile clients properly, their
name will be derived from it.

Note we can't really use the <symbol category="filename" name="...">
element from ycpc's XML to determine the file name because the parser
always passes a temporary file with "unique" file name to ycpc, not the
original file. This is something I don't want to change (mostly to
ensure multiple instances of y2r can be run concurrently safely while
keeping the code simple).