Fix Client to be insise YCP::Clients module and have Client suffix
Reason for YCP::Clients module is that this namespace does not exist,
so ruby raise unknown constant.
Client suffix is needed because in ycp is often client which have same name as
module like sysconfig.ycp and Sysconfig.ycp with module and when Sysconfig
module is called, then wrong class is used.
Reason why commit is not split into two is that tests have so much repetition
of client code, that change it even once was quite lot of work.