Translate the |switch| statement
For now we only translate |switch| statements where each case has a
|break| statements at the end. In other words, we don't implement
translation of |switch| statenents with overlapping cases. This is done
for simplicity (there is no direct equivalent for |switch| statements
  with overlapping cases in Ruby).

We may revise the decision in future if we find enough switch statements
with overlapping cases in YaST codebase.