Unify function and variable handling somewhat
We want to use |ruby_var| to escape names of nested functions in the
same way as we do with variables, because they are actually stored in
variables in genereated Ruby code. To make this work, |ruby_var| needs
to be aware of surrounding variables *and functions*. This commit makes
it so.

Mixing function and variables together like this is OK, because there
can't be both a function and a variable of the same name defined at the
same block.