Specs: Use "i", "j" and "k" variables in assignments
This makes the variables different from "a", "b" and "c" which are often
used as arguments of functions or blocks. This ensures that assignemnts
used as function or block bodies will translate the same independent on
the context and passed/non-passed argument.

This is not immediately needed, but it will be in the near future.