Handle function argument redefinition
YCP allows to define a variable with the same name as a function
argument inside a function. Without this commit, we created a duplicate
entry in the symbol table in this case, which lead to "2" suffix
appended to the variable name in generated Ruby code. It could have also
caused some other problems as Y2R code generally assumes unique entries
in symbol tables.