Introduce the Parser::Context class
This class is similar to AST::YCP::Context in that it transfers context
between method calls walking the tree (in this case, Nokogiri's XML
representation). It is needed because we'll need to transfer more
context soon (see the next commits), so a simple symbol, which was used
until now, would not suffice.