Translate |textdomain| without |include|
Before this commit, the |textdomain| statement was translated as:

  include I18n
  textdomain "foo"

This caused problems when the |textdomain| statement was used inside a
function -- see #10.

This commit changes the translation to:

  textdomain "foo"

The |include| is not needed anymore as all clients & modules include
I18n by default (because it is included in their base class).

Fixes #10.