Fix compilation of |if| statements with no then
An <if> element inside ycpc-produced XML corresponding to an |if|
statement with an empty block in the "then" part won't contain a <then>
subelement. Because our code relied on its presence, it compiled such
statements incorrectly (it trasnlated an "else" part of such |if|
statement as its "then" part). This commit fixes the problem.