Formatting: Improve computation of available space
Before this change, context sent to #to_ruby usually contained only the
|width| property which contained the number of remaining characters
until the margin (hardcoded as 80 for now) *at the current line*. This
caused trouble e.g. in cases like this:

  1                        80
  |---------- ... ----------|

  LongModule.long_method([
    [
      item1,
      item2,
      item3
    ]
  ])

Although the nested array containing |item1|, |item2| and |item3| would
fit into one line, it was split into multiple lines because the passed
|width| was only 3 (the number of characters after the first |[| until
the margin.

This commit splits the original |width| property into |width| and
|shift|. The width has tha same semantics as it didi before, but it only
cares about the indent level, not about the code that is already on the
current line. The width of that code is stored in |shift|. As a result,
each #to_ruby method knows how much available space there is on the
*current line* and also on the *next lines*.