Compile clients without the YCP::Clients namespace
So far, classes of compiled clients were put into the YCP::Clients
namespace. This is not really needed as their names already contain a
"Client" suffix, which is enough to prevent conflicts with other thing
in the main YCP namespace.

This commit removes the compilation into the YCP::Clients namespace.