Pass options[:as_include_file] to the parser
Originally I thought I'll deal with special compilation of includes on
the YCP AST level. Now I think that I'll be able to deal with include
files completely on the parser level. This means the information whether
we are compiling an include file is needed there. This commit adds it.