Optionally compile |include| statements as |YCP.include|
Until now we let ycpc to inline include files for us. This commit
introduces capability to remove inlined include files code and instead
compile the |include| statements as |YCP.include|.

This capability is hidden behind a command-line option
(--dont-inline-include-files), but this is just temporary. Once we make
sure this works properly and the Ruby bindings implement |YCP.include|,
we will probably remove the option and just always inline.

Note the uninlining works only for toplevel includes. The plan is to
make Y2R detect non-toplevel includes later and fix these manually.

Parts of this commit are based on a patch by Josef Reidinger.