Fix <include> and <filename> handling with --extract-file
Before this commit, all AST nodes representing <include> and <filename>
elements were discarded when --extract-file was used. This breaks
--dont-inline-include-files because it relies on these nodes to detect
include file boundaries. This manifested by inlining nested includes
when both --extract-file (on the outer include) and
--dont-inline-include-files were used.

This commit fixes the code to keep the AST nodes which are part of the
extracted statements intact.

Based on a patch by Josef Reidinger.