Add the --report-file option
The --report-file option allows YCP Killer to specify the name of the
real translated file with full YaST module-relative path.

This information is needed when --extract-file is used. In that case,
the input file is include wrapper file, not the include itself, and the
value passed to the --extract-file option is relative to the exported
directory, not YaST module root.

Part of a fix for #23.