Fix a bug in multi-line comment handling
The part of the regexp responsible for parsing the inside of multi-line
comments contained a bug. If there was a star ("*"), the regexp always
consumed the character after it. This caused a problem for comments like
this:

  /***/

Here the rule consumed stars #2 and #3 as part of the inside of the
comment. As a result, the ending delimiter wasn't found and the comment
wasn't handled properly.

This commit fixes the problem by using a look-ahead expression.