Fix regexps inside Comments.fix_delimiters
The regexps didn't match only at the beginning of the segment, but also
elsewehere. This lead to wrong handling of comments like this:

  /*
  // a
  */

Here the /^\/\// regexp matched and the comment was threated as a
one-line slash comment, not as a multi-line comment. This resulted into
meitthing the followign comment into Ruby code:

  /*
  # a
  */

This comment fixes the regexps to match at the beginning of the segment
only.