Ruby AST: s/no_comments/base/
Nodes are generally supposed to define #to_ruby_no_comments method,
which emits core of the Ruby code of the node. The more general #to_ruby
method then mechanically wraps this code with comments, if the node has
any.

This commit changes the name from #to_ruby_no_comments to #to_ruby_base.
This is because #to_ruby may not do only comment-processing in the
future and the old name wouldn't be accurate in that case. The new name
is also a bit shorter.

Similar reasoning holds for #single_line_width_without_comments, which
gets renamed to #single_line_width_base.