Comments: Fix whitespace handling
Before this change, the following code:

  y2milestone("M1");

  y2milestone("M2");

was translated into:

  y2milestone("M1");
  y2milestone("M2");

The empty line between the calls was missing. This happened because
|process_comment_before| produced an empty string, which was ignored by
the caller.

This commit adjusts the code to no ignore empty strings. If a comment
really should be ignored, |process_comment_{after,before}| returns
|nil|.

This commit also causes empty lines sometimes to appear at the beginning
of emitted methods or modules. This will be addressed later.