Ruby AST: Fix emitting of |next| with before-commented value
Before this commit, |next| statements with a value which had comment
before were emitted like this:

  next
  # before
  42

This was wrong, because Ruby saw |next| and |42| as two separate
statements.

This commit fixes the emitting to something this:

  next (
    # before
    42
  )

The parens are ugly, but it was the only solution which didn't require
extensive movement of comments around nodes.