Comments: Remove |comment_after| hack in the parser
The hack is not needed now, because the new whitespace processing code
already handles the situation the hack tries to prevent.

Trailing newlines, which the hack tries to remove, get attached as
|comment_after| to some node, which is always the last node of some node
sequence (possibly one-element). But for these, we always call
|Comments.process_comment_after| with options set to remove whitespace
below |comment_after|.

Also note the condition in the hack was wrong. It removed all
after-comments *containing an whitespace-only line*, not *consisting
only of whitespace-only line*.