Handle starred comments with non-empty first line better
Before this commit we had problems handling starred comments with
non-empty first line, e.g.:

  /* line 1
   * line 2
   * line 3
   */

Our starring detection (detection of the vertical line of stars) was
performed only after removal of initial and terminal delimiters with
associated whitespace. At that point, the comment looked like this:

  line 1
  * line 2
  * line 3

The starring detection failed at that point because not all lines were
prefixed with stars. This meant the comment was not processed correctly
and was emitted in Ruby in this (wrong) form:

  # line 1
  # * line 2
  # * line 3

This commit moves the starring detection code to earlier stage and
modifies the delimiter removing code to handle starred comments
correctly. As a result, the comment above is now emitted in this
(correct) form:

  # line 1
  # line 2
  # line 3