Revert "do not emit nil for functions that return void"
It turned out that it is possible to write e.g.

  void v = f();

where |f| is a |void| function. If we translate this code to Ruby and
optimize away trailing |nil| in |f|, |v| might get an invalid value.
Something like this actually happened in testsuites of some modules
(e.g. yast2).

This reverts commit 47640f222b8d65e33f32389f1fe5d053e60d0bdf.

Conflicts:
	lib/y2r/ast/ycp.rb