Ruby AST: Make decisions about enclosing into parens priority-based
Before this commit, eclosing into parens was a hack. For certain kinds
of expressions (e.g. binary operators), we basically eclosed their
operands whenever they were something non-trivial. This lead to ugly
code e.g. in cases where multiple operators with the same priority were
combined or in common cases like this:

  if (foo != nil) && (foo != false)
    ...
  end

This commit changes the mechanism to a priority-based one. When emitting
code, the priority in current context is known and parens are emitted
only when the expression emitted has lower priority than the priority in
current context. In other words, we emit parens only when they are
strictly necessary.