Fix skipped includes emitting when using --extract-file
When using --extract-file, skipped includes were not emitted in the Ruby
code. This caused a problems e.g. in the following situation:

  * client.ycp includes include1.ycp and include2.ycp
  * include2.ycp includes include1.ycp
  * Y2R is executed on client.ycp with --extract-file include2.ycp

In this situation, the <include name="include1.ycp"> element in
include2.ycp was marked with skipped="1" (because include1.ycp was
already included in client.ycp) and it wasn't emitted in Ruby code. If
include2.ycp was included in some client that didn't include
include1.ycp before including include2.ycp, we were in deep shit.

This commit makes sure that even skipped includes are emitted. We rely
on Ruby bindings to detect that an include is included for the second
time and not execute its initialization again.

Should fix https://github.com/yast/ycp-killer/issues/559.